requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.11
django-ckeditor==5.4.0
django-js-asset==1.1.0
djangorestframework==3.8.2
docopt==0.6.2
idna==2.6
Pillow==5.1.0
pipreqs==0.4.9
pytz==2018.4
requests==2.18.4
urllib3==1.22
yarg==0.1.9
